+++ docs/components/example/checkbox-stronger.tsx
@@ -1,5 +1,0 @@
-import { Checkbox } from "seed-design/ui/checkbox";
-
-export default function CheckboxStronger() {
-  return <Checkbox label="Bold Label Text" weight="stronger" />;
-}