+++ packages/figma/src/entities/data/__generated__/component-sets/index.d.ts
@@ -21,6 +21,8 @@ export { metadata as contextualFloatingButton } from "./contextual-floating-butt
 export { metadata as divider } from "./divider";
 export { metadata as floatingActionButton } from "./floating-action-button";
 export { metadata as helpBubble } from "./help-bubble";
+export { metadata as listHeader } from "./list-header";
+export { metadata as listItem } from "./list-item";
 export { metadata as mainTabNavigationGlobal } from "./main-tab-navigation-global";
 export { metadata as mainTabNavigationKr } from "./main-tab-navigation-kr";
 export { metadata as mannerTemp } from "./manner-temp";