+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @seed-design/mcp
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [afdd1ee]
+  - @seed-design/figma@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes